src: move CallbackScope to separate cc/h
