+++ src/node.h
@@ -63,6 +63,7 @@
 #include "v8.h"  // NOLINT(build/include_order)
 #include "v8-platform.h"  // NOLINT(build/include_order)
 #include "node_version.h"  // NODE_MODULE_VERSION
+#include "callback_scope.h"
 
 #define NODE_MAKE_VERSION(major, minor, patch)                                \
   ((major) * 0x1000 + (minor) * 0x100 + (patch))
@@ -593,12 +594,6 @@ typedef void (*promise_hook_func) (v8::PromiseHookType type,
                                    v8::Local<v8::Value> parent,
                                    void* arg);
 
-typedef double async_id;
-struct async_context {
-  ::node::async_id async_id;
-  ::node::async_id trigger_async_id;
-};
-
 /* Registers an additional v8::PromiseHook wrapper. This API exists because V8
  * itself supports only a single PromiseHook. */
 NODE_EXTERN void AddPromiseHook(v8::Isolate* isolate,
@@ -646,36 +641,6 @@ NODE_EXTERN async_context EmitAsyncInit(v8::Isolate* isolate,
 /* Emit the destroy() callback. */
 NODE_EXTERN void EmitAsyncDestroy(v8::Isolate* isolate,
                                   async_context asyncContext);
-
-class InternalCallbackScope;
-
-/* This class works like `MakeCallback()` in that it sets up a specific
- * asyncContext as the current one and informs the async_hooks and domains
- * modules that this context is currently active.
- *
- * `MakeCallback()` is a wrapper around this class as well as
- * `Function::Call()`. Either one of these mechanisms needs to be used for
- * top-level calls into JavaScript (i.e. without any existing JS stack).
- *
- * This object should be stack-allocated to ensure that it is contained in a
- * valid HandleScope.
- */
-class NODE_EXTERN CallbackScope {
- public:
-  CallbackScope(v8::Isolate* isolate,
-                v8::Local<v8::Object> resource,
-                async_context asyncContext);
-  ~CallbackScope();
-
- private:
-  InternalCallbackScope* private_;
-  v8::TryCatch try_catch_;
-
-  void operator=(const CallbackScope&) = delete;
-  void operator=(CallbackScope&&) = delete;
-  CallbackScope(const CallbackScope&) = delete;
-  CallbackScope(CallbackScope&&) = delete;
-};
 
 /* An API specific to emit before/after callbacks is unnecessary because
  * MakeCallback will automatically call them for you.
